MCS-274 Moved scene generator find_performer_rect function from util to geometry class.
